Remove bigquery dataset feature
I'd like to rethink the strategy around persisting generated scores and
persisting results in general. Especially now that we've removed the
requirement on BigQuery from the main path.

For a first release, I don't think we need this functionality.

We can always adapt it back in from git history later down the line.