Improve progress logging
Generating scores for lots of repositories can take a long time and it
wasn't apparent before from the output how many scores needed generating
or how far through the list you were.

I've added a progress bar that tracks how many repositories we've found
and where we are in the process of fetching scores from them.

We also print a message that indicates what tally is up to, i.e:
"Generating score for 'X'".

As part of this I've created a scorecard.Client interface that allows us
to iterate through different methods for fetching scorecard results
easily. We try every repository with the first client, then move onto the next
client after the first has had a go at all the repos.

This should make it simpler to add new sources of scorecard data in
future.

Doing this work made me realise that our output format returns a result
for every package+(repository/score) combination, when really what I think we
want is a result per-score (to avoid multiple results with the same
score). So I've updated the output format to reflect that.

I updated to Go 1.20 while working on this because I wanted the new
error wrapping capabilities of errors.Join(). This is because the errors
from the scorecard generation are difficult to handle, so I want to wrap
every error as 'not found' so we know we can ignore the failure.