Add a local cache for retrieved scores
Fetching scores from the API can take a bit of time, so it would be nice to
cache the response to speed up subsequent runs of tally.

This implements a cache that stores scores in a local sqlite database.

There's a scorecard.Client implementation that wraps around another
client, transparently trying to retrieve scores from the cache before
falling back to the client.

This will also cache generated scores, which is a boon, considering how
long they can take to generate.

The database is stored in `~/.cache/tally/cache/cache.db` by default.
This can be changed with the `--cache-dir` flag.

Scores are cached for 7 days, which is the general cadence of scorecard
runs. This can be tweaked with the `--cache-duration` flag.

Caching can be disabled altogether with `--cache=false`.