Fetch scores from the scorecard API
Fetching scores from our own database has some limitations and problems as an
approach.

1. We can't store all the information from a scorecard result, or the
   database would be far too large.
2. The information in the database will invariably be stale to a lesser
   or greater degree.
3. At some point Scorecard are going to add support for non-Github
   repositories and this is going to add significantly to the database
   size.
4. It requires that we run infrastructure and take on the operational
   burden of generating the database. We're ultimately responsible for
   ensuring this keeps working.

Instead, lets fetch scores from the API. This will take a bit longer,
depending on the size of the BOM but its ultimately more scalable and I
think the time taken to go to the API is basically 'the cost of doing
business'. It will also allow us to capture the full scorecard result
information.

To speed things up, I've started fetching scores concurrently, where the
client implementation supports it. Currently it doesn't seem like you
can run the scorecard client concurrently. I need to investigate this
further.

In future we could implement a cache to make this a nicer experience.

This commit doesn't remove the DB code. I'll do that in a follow-up.