Support any package type
Initially we only supported the types in the deps.dev database. However,
now that we can get repository names from the BOM itself, there's no
reason we should be ignoring other types.

As part of this I've modified the way that we represent packages so that
it more closely matches the Purl fields, rather than the fields in
deps.dev.